refactor：OneFragmentにあったdiff_utilをCustomAdapterへ移動
